requirements.txt
# Core
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.0
python-dotenv>=1.0.0

# LangChain AI layer (optional — only needed if USE_LANGCHAIN=true)
langchain>=0.2.0
langchain-anthropic>=0.1.0
langchain-core>=0.2.0
anthropic>=0.25.0
